Add e2e test checking if next page button works
This is a more complex commit compared to the previous few commits.

At a high level, all it does is add an end-to-end test verifying that clicking on the next page button shows the next page of search results.

To accomplish this, a few changes to newly created files as well as previously existing ones were necessary.

The creation of search-results-page-2.json is necessary to mock a second page of search results for Cypress. I created the contents of that file by copying the URL sent by this application when clicking on the network tab of the browser's development tools after clicking for the next page of search results when on the first page of search results for search query "hello".

Then, I pasted that URL into a browser window while authenticated to see what the response looks like in JavaScript object notation, or .json, and copied and pasted the contents of that file into search-results-page-2.json.

In addition, I had to modify one of the previously existing intercepts to not catch all routes to search queries, since we want a different fixture to appear when rendering the second page of search results to confirm that a change occurred.

The changes to Paginate.tsx are trivial: I just gave the first page button and next page button end-to-end ids so that Cypress could access them while testing.

The changes in app_spec.ts involve the most lines of code. It sets up a describe block to lay the groundwork for testing of the pagination element. It visits a page that would have the pagination bar, like the page that appears when entering the search query "hello", and checks, first, that the expected first page of search results is present.

Then, it checks that the first and next page buttons are visible before clicking the next page button.

It needs to wait about two seconds to give the DOM enough time to render the second page of search results.

Then, it checks that the second page of search results appear as expected. Hope that helps!