add pagination (#290)
* Lay groundwork for pagination to close #169

* Remove border from top of pagination bar

* Remove bottom padding to fix spacing below results

* Change results per page from 5 to 10 as requested

* Fix borders of page buttons on hover and selection

* Auto-fix spacing of app_spec.ts; no manual changes

I did not make any manual changes in this commit. This commit automatically fixes the spacing in `app_spec.ts`. I did this by saving `app_spec.ts` twice. I didn't add anything to the body of the file.

* Rename variable denoting results per page to limit

This is a minor infrastructural change. There should be no changes in behavior.

All this does is rename the variable denoting the number of results per page from `page_size` to `limit` to match JavaScript's camelcase convention.

After doing that, I also removed the `// eslint-disable-next-line camelcase` comments because they were no longer necessary. Hope it helps!

* Improve naming of function that finds page count

This is another minor infrastructural change.

This change shouldn't cause any changes in this application's behavior.

All it does is rename findNumberOfPages to getPageCount because findNumberOfPages implies use of the .find() function.

* Add e2e test checking if next page button works

This is a more complex commit compared to the previous few commits.

At a high level, all it does is add an end-to-end test verifying that clicking on the next page button shows the next page of search results.

To accomplish this, a few changes to newly created files as well as previously existing ones were necessary.

The creation of search-results-page-2.json is necessary to mock a second page of search results for Cypress. I created the contents of that file by copying the URL sent by this application when clicking on the network tab of the browser's development tools after clicking for the next page of search results when on the first page of search results for search query "hello".

Then, I pasted that URL into a browser window while authenticated to see what the response looks like in JavaScript object notation, or .json, and copied and pasted the contents of that file into search-results-page-2.json.

In addition, I had to modify one of the previously existing intercepts to not catch all routes to search queries, since we want a different fixture to appear when rendering the second page of search results to confirm that a change occurred.

The changes to Paginate.tsx are trivial: I just gave the first page button and next page button end-to-end ids so that Cypress could access them while testing.

The changes in app_spec.ts involve the most lines of code. It sets up a describe block to lay the groundwork for testing of the pagination element. It visits a page that would have the pagination bar, like the page that appears when entering the search query "hello", and checks, first, that the expected first page of search results is present.

Then, it checks that the first and next page buttons are visible before clicking the next page button.

It needs to wait about two seconds to give the DOM enough time to render the second page of search results.

Then, it checks that the second page of search results appear as expected. Hope that helps!

* Give e2e test more time to show new search results

Tests are failing on push but passing locally. I think this is because whatever machine is running the test needs more time to display updated search results, so I'm telling Cypress to wait for 5 seconds instead of 2 before continuing with the pagination element's test.

* Capitalize interface name to follow TS convention

* Give e2e test even more time to show new results

Tests still fail on a remote server even while passing locally, so I'm going to give the server running the pagination element test a whopping 30000 milliseconds, or 30 seconds, to wait for the second page of search results to render before continuing with the pagination element test.

* Make e2e test wait for second page of results

Instead of having the e2e pagination test wait 30 seconds before testing the second page of search results, it waits for the second page of search results to render, continuing with testing immediately.

* chore: format cypress fixtures

* Change prop type to avoid antipattern

* Specify stricter typing for onPageClick

* Use shorthand and remove duplicate context in name

* Remove duplicate context in state variable name

* Remove naming redundancy in test for pagination

* Prefix test name to follow testing convention

* fix: fix failing e2e tests

* fix: temporarily skip failing pagination e2e test

* Use `&&` operator instead of unnecessary ternaries

* Make correction in copied-and-pasted comment

* Fix a couple of code comment typos

* Update code comments to reflect change to ternary

* Remove border-gray styling that didn't do anything

* Remove text-gray styling that didn't do anything

* Simplify classNames by moving text-sm to parent

* Remove "font-medium" since it doesn't do anything

* Remove unnecessary "bg-primary"s to reduce clutter

* Remove "relative" since it doesn't do anything

* Remove className that doesn't cause any UI changes

* Remove pagination bar's unnecessary enclosing div

* Remove pagination text's unnecessary enclosing div

* Remove hover styling from unclickable ellipsis

* Removing redundant "text-secondary" className

* Fix plural to handle edge case of exactly 1 result

* Only show pagination bar for >= 2 pages of results

* Keep next button on the right on page 1 for mobile

* Fix conceptual code comment error

* Handle edge case of zero results

* Use Ionicons for back and forward to simplify code

* Add nav buttons for jumping to first and last page

* Make navigation buttons playback-themed!

* Revert "Make navigation buttons playback-themed!"

This reverts commit ea90eba95e6183acf2c8a52c7ce8c661caf394d7.

* Revert "Add nav buttons for jumping to first and last page"

This reverts commit ba9a03c44058bf90ba7bb275460d4b2d892b9e28.

* Add screen reader text for increased accessibility

* Shorten name of function to improve readability

* Shorten name of variable for improved readability

* Make fn variable names shorter and more consistent

* Always show current page, even if in a middle page

* Make test name more concise by removing a word

Co-authored-by: Amila Welihinda <amilajack@gmail.com>